changed settings ui to use TimerInput to select fade duration for sleep timer

+++ app/src/main/java/ie/delilahsthings/soothingloop/MainActivity.java
@@ -551,6 +551,6 @@ public class MainActivity extends AppCompatActivity {
 
         populateNoiselist();
         populateCustomNoiselist();
-        SoundEffectVolumeManager.fadeOut(this, settings.getInt(Constants.FADEOUT_DURATION, 3)*1000);
+        SoundEffectVolumeManager.fadeOut(this, settings.getLong(Constants.FADEOUT_DURATION, 3)*1000);
     }
 }
